# docs/openapi.json
{
  "components": {
    "schemas": {
      "AgentName": {
        "enum": [
          "orchestrator",
          "extractor",
          "compliance",
          "auditor",
          "corrective",
          "reviewer",
          "system"
        ],
        "title": "AgentName",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EventType": {
        "enum": [
          "WORKFLOW_CREATED",
          "DOCUMENT_RECEIVED",
          "EXTRACTION_STARTED",
          "EXTRACTION_COMPLETED",
          "EXTRACTION_FAILED",
          "COMPLIANCE_CHECK_STARTED",
          "FINDING_CREATED",
          "AUDIT_STARTED",
          "AUDIT_CHECK_RESULT",
          "AUDIT_CONFLICT_FOUND",
          "REEXTRACTION_REQUESTED",
          "REEXTRACTION_COMPLETED",
          "AUDIT_COMPLETED",
          "HUMAN_REVIEW_REQUIRED",
          "CORRECTIVE_ACTION_DRAFTED",
          "REVIEW_DECISION_RECORDED",
          "RULES_RERUN_STARTED",
          "RULES_RERUN_COMPLETED",
          "DISPATCH_RECORDED",
          "WORKFLOW_COMPLETED",
          "WORKFLOW_FAILED"
        ],
        "title": "EventType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoredEvent": {
        "properties": {
          "agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentName"
              },
              {
                "type": "string"
              }
            ],
            "title": "Agent"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "finding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow_id",
          "seq",
          "type",
          "agent",
          "timestamp",
          "message"
        ],
        "title": "StoredEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowStatus": {
        "enum": [
          "CREATED",
          "EXTRACTING",
          "EVALUATING",
          "AUDITING",
          "AWAITING_REVIEW",
          "AWAITING_EVIDENCE",
          "COMPLETED",
          "FAILED"
        ],
        "title": "WorkflowStatus",
        "type": "string"
      },
      "WorkflowSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "document_count": {
            "default": 0,
            "title": "Document Count",
            "type": "integer"
          },
          "finding_count": {
            "default": 0,
            "title": "Finding Count",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "needs_review_count": {
            "default": 0,
            "title": "Needs Review Count",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "rule_set_id": {
            "default": "",
            "title": "Rule Set Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "status",
          "created_at",
          "updated_at",
          "provider",
          "model"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Evidence reconciliation with deterministic rules, an independent Auditor and human approval. Demo Buyer Framework on Synthetic Demo Data.",
    "title": "ChainAudit.ai API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workflows": {
      "get": {
        "operationId": "list_workflows_api_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowSummary"
                  },
                  "title": "Response List Workflows Api Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/workflows/{workflow_id}/events": {
      "get": {
        "description": "Server-Sent Events: stored history after `Last-Event-ID`/`after`, then live events.",
        "operationId": "stream_events_api_workflows__workflow_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "description": "Replay events with seq > after",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replay events with seq > after",
              "title": "After"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "events"
        ]
      }
    },
    "/api/workflows/{workflow_id}/events/history": {
      "get": {
        "operationId": "event_history_api_workflows__workflow_id__events_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StoredEvent"
                  },
                  "title": "Response Event History Api Workflows  Workflow Id  Events History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event History",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}